🤡 issues 데이터 내부의 마일스톤 정보를 MILESTONE_TABLE로 변경 - import 로 MILESTONE_TABLE을 불러오면 undefind인 문제가 생겨서 table을 issue.ts로 이동함
