+++ src/mocks/handlers/issue.ts
@@ -1,12 +1,11 @@
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { rest } from 'msw';
-import { issues, issueTable } from '@/mocks/tables/issue';
+import { issues, issueTable, MILESTONE_TABLE } from '@/mocks/tables/issue';
 import { REACTIONS } from '@/components/Molecules/Dropdown/Panel/Reaction/mock';
 import { IssuesTypes, CommentsTypes, ContentTypes, IssueHistoryTypes, ReactionResponseTypes } from '@/api/issue/types';
 import { filterIdPassword, TEST_USER, USER_TABLE } from '@/mocks/handlers/auth';
 import { responseNewIssueData } from '@/mocks/tables/newIssueHelper';
 import { LABEL_TABLE } from '@/mocks/handlers/label';
-import { MILESTONE_TABLE } from '@/mocks/tables/milestone';
 import { findMilestoneHelper } from '@/mocks/handlers/milestone';
 import {
   assigneesHistory,
